+++ .github/workflows/release.yml
@@ -89,6 +89,15 @@ jobs:
         if: matrix.build == 'macos'
         run: brew install p7zip
 
+      - name: Install vcpkg OpenSSL (Windows)
+        if: matrix.build == 'windows'
+        run: |
+          git clone https://github.com/microsoft/vcpkg
+          .\vcpkg\bootstrap-vcpkg.bat
+          .\vcpkg\vcpkg install openssl:x64-windows
+        env:
+            VCPKGRS_DYNAMIC: 1
+
       - name: Build (Linux)
         if: matrix.build == 'linux'
         run: |
@@ -104,6 +113,7 @@ jobs:
         run: cargo build --release
         env:
           RUSTFLAGS: -Ctarget-feature=+crt-static
+          VCPKGRS_DYNAMIC: 1
 
       - name: Create artifact directory
         run: |
