+++ Cargo.toml
@@ -37,7 +37,7 @@ lazy_static = "1.4.0"
 log = "0.4.11"
 notify = "4.0.15"
 number_prefix = "0.4.0"
-openssl = { version = '0.10.29', optional = true }
+openssl = { version = "0.10.29"}
 percent-encoding = "2.1.0"
 predicates = "1.0.5"
 prettytable-rs = "0.8.0"
