+++ src/terminal/mod.rs
@@ -1,4 +1,6 @@
+mod browser;
 pub mod emoji;
 pub mod interactive;
 pub mod message;
 pub mod styles;
+pub use browser::open_browser;
